relocated files
 To avoid confusions files were relocated on new directories that better depict their use

- handling undefined token error
