+++ src/App.jsx
@@ -1,10 +1,10 @@
 import React, { useState, useEffect } from 'react';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
 //
-import Navbar from './components/Navbar.jsx';
-import LoginPage from './components/LoginPage.jsx';
-import HomePage from './components/HomePage.jsx';
-import RecipeEditor from './components/RecipeEditor.jsx';
+import Navbar from './components/navbar/Navbar.jsx';
+import LoginPage from './components/logIn/LoginPage.jsx';
+import HomePage from './components/mainPages/HomePage.jsx';
+import RecipeEditor from './components/mainPages/RecipeEditor.jsx';
 //
 import {
   getFromLocalStorage,
@@ -17,7 +17,9 @@ function App() {
 
   useEffect(() => {
     // Store the token every time is updated
-    saveOnLocalStorage('token', token);
+    if (token !== undefined) {
+      saveOnLocalStorage('token', token);
+    }
   }, [token]);
 
   // If there's no token take the user back to the log in page
